Ajout de la confirmation de suppression d'inscription+correction bug ajout de groupe

+++ GestAssoc/GestAssoc.Modules.GestionInscriptions/Commands/DeleteInscriptionCommand.cs
@@ -1,9 +1,8 @@
-﻿using GestAssoc.Common.Commands;
-using GestAssoc.Common.Utility;
+﻿using GestAssoc.Common.Utility;
 using GestAssoc.Model.Models;
-using GestAssoc.Modules.GestionInscriptions.Constantes;
 using GestAssoc.Modules.GestionInscriptions.Properties;
 using GestAssoc.Modules.GestionInscriptions.Services;
+using Microsoft.Practices.Prism.Interactivity.InteractionRequest;
 using Microsoft.Practices.ServiceLocation;
 using Microsoft.Practices.Unity;
 using System;
@@ -13,6 +12,14 @@ namespace GestAssoc.Modules.GestionInscriptions.Commands
 {
 	public class DeleteInscriptionCommand : ICommand
 	{
+		public InteractionRequest<IConfirmation> RqConfirmDelete { get; private set; }
+		private Action _commandCallBack;
+
+		public DeleteInscriptionCommand(Action callback) {
+			this.RqConfirmDelete = new InteractionRequest<IConfirmation>();
+			this._commandCallBack = callback;
+		}
+		
 		public bool CanExecute(object parameter) {
 			return true;
 		}
@@ -23,22 +30,33 @@ namespace GestAssoc.Modules.GestionInscriptions.Commands
 		}
 
 		public void Execute(object parameter) {
-			var service = ServiceLocator
-				.Current.GetInstance<IUnityContainer>()
-				.Resolve<IGestionInscriptionsServices>();
+			this.RqConfirmDelete.Raise(
+				new Confirmation
+				{
+					Content = Resources.Confirm_SuppressionInscription + Environment.NewLine + (parameter as Inscription).ToString(),
+					Title = Common.Properties.Resources.Titre_Confirmation
+				},
+				c => this.ExecuteCallback(c.Confirmed, parameter as Inscription)
+			);
+		}
 
-			var itemToDelete = parameter as Inscription;
+		private void ExecuteCallback(bool deleteConfirmed, Inscription itemToDelete) {
+			if (deleteConfirmed) {
+				var service = ServiceLocator
+					.Current.GetInstance<IUnityContainer>()
+					.Resolve<IGestionInscriptionsServices>();
 
-			try {
-				UIServices.SetBusyState();
-				service.DeleteInscription(itemToDelete);
+				try {
+					UIServices.SetBusyState();
+					service.DeleteInscription(itemToDelete);
 
-				NotificationHelper.WriteLog(Resources.Log_EnregistrementSupprime);
+					NotificationHelper.WriteLog(Resources.Log_EnregistrementSupprime);
 
-				new ShowViewCommand(ViewNames.ConsultationInscriptions.ToString()).Execute(null);
-			}
-			catch (Exception ex) {
-				NotificationHelper.ShowError(ex);
+					this._commandCallBack();
+				}
+				catch (Exception ex) {
+					NotificationHelper.ShowError(ex);
+				}
 			}
 		}
 	}
